+++ shardis-ui/config/spec-bundle.js
@@ -14,6 +14,7 @@ require('ts-helpers');
 require('zone.js/dist/zone');
 require('zone.js/dist/long-stack-trace-zone');
 require('zone.js/dist/jasmine-patch');
+require('zone.js/dist/async-test');
 
 require('rxjs/Rx');
 require('rxjs/subject/ReplaySubject');