+++ shardis-ui/src/main/frontend/app/views/home/home.spec.ts
@@ -1,7 +1,7 @@
 import {
   it,
   inject,
-  injectAsync,
+  async,
   describe,
   beforeEachProviders,
   TestComponentBuilder