+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * fully stateless authentication
 * storing oauth2 token in localstorage
 * removed form authentication
+* removed deprecated injectAsync
+
 
 ### 1.3.0 (21.07.2016)
 * refactoring to microservices